modal btn 로직 구현 완료
