+++ src/components/Login/LoginForm.jsx
@@ -42,6 +42,7 @@ const LoginForm = () => {
     };
 
     const success = await loginHook(loginData);
+    console.log(success);
     if (success.token) {
       Cookies.set('token', success.token.split('.')[1]);
       localStorage.setItem('user_id', success.user_id);
