+++ src/styles/items/menu/dropmenu.module.css
@@ -1,7 +1,6 @@
 .menu {
   position: absolute;
   background-color: white;
-
   display: flex;
   flex-direction: column;
 }
