+++ src/styles/items/modal/non_login_users_modal.module.css
@@ -4,6 +4,7 @@
   left: 0;
   width: 100%;
   height: 100%;
+  z-index: 99;
 }
 
 .modal_overlay {
@@ -38,6 +39,7 @@
   font-weight: 900;
   line-height: 3rem;
   margin-bottom: -10px;
+  color: black;
 }
 
 .to_login_button {
